refactor vectorstore
